+++ python/sglang/srt/layers/utils/logprob.py
@@ -1,0 +1,24 @@
+from typing import List
+
+import torch
+
+
+def get_top_logprobs(
+    logprobs: torch.Tensor,
+    top_logprobs_nums: List[int],
+):
+    max_k = max(top_logprobs_nums)
+    ret = logprobs.topk(max_k, dim=1)
+    values = ret.values.tolist()
+    indices = ret.indices.tolist()
+
+    output_top_logprobs_val = []
+    output_top_logprobs_idx = []
+    for i, k in enumerate(top_logprobs_nums):
+        output_top_logprobs_val.append(values[i][:k])
+        output_top_logprobs_idx.append(indices[i][:k])
+
+    return (
+        output_top_logprobs_val,
+        output_top_logprobs_idx,
+    )
